+++ src/xSTIR/pSTIR/STIR.py
@@ -4,7 +4,7 @@ Object-Oriented wrap for the cSTIR-to-Python interface pystir.py
 
 # CCP PETMR Synergistic Image Reconstruction Framework (SIRF)
 # Copyright 2015 - 2017 Rutherford Appleton Laboratory STFC
-# Copyright 2015 - 2017 University College London
+# Copyright 2015 - 2017, 2019 University College London
 #
 # This is software developed for the Collaborative Computational
 # Project in Positron Emission Tomography and Magnetic Resonance imaging
@@ -559,6 +559,68 @@ class RayTracingMatrix(object):
         '''
         return parms.int_par(self.handle, self.name, 'num_tangential_LORs')
         #return _int_par(self.handle, self.name, 'num_tangential_LORs')
+
+class SPECTUBMatrix:
+    '''
+    Class for objects holding sparse matrix representation of a SPECT
+    projector (developed at the University of Barcelona) (see AcquisitionModel class).
+    '''
+    name = 'SPECTUBMatrix'
+
+    def __init__(self):
+        '''
+        Create a new matrix. Default settings use neither attenuation nor resolution modelling.
+        '''
+        self.handle = pystir.cSTIR_newObject(self.name)
+        check_status(self.handle)
+
+    def __del__(self):
+        if self.handle is not None:
+            pyiutil.deleteDataHandle(self.handle)
+    def set_keep_all_views_in_cache(self, value):
+        '''
+        Enable keeping the matrix in memory.
+
+        This speeds-up the calculations, but can use a lot of memory.
+
+        You have to call set_up() after this (unless the value didn't change).
+        '''
+        _set_int_par(self.handle, self.name, 'keep_all_views_in_cache', value)
+        return self
+    def get_keep_all_views_in_cache(self):
+        '''
+        Returns a bool checking if we're keeping the whole matrix in memory or not.
+        '''
+        return _int_par(self.handle, self.name, 'keep_all_views_in_cache') != 0
+    def set_attenuation_image(self, value):
+        '''
+        Sets the attenuation image used by the projector.
+        '''
+        assert_validity(value, ImageData)
+        _setParameter(self.handle, self.name, 'attenuation_image', value.handle)
+        return self
+    def get_attenuation_image(self):
+        '''
+        Returns the attenuation image used by the projector.
+        '''
+        image = ImageData()
+        image.handle = _getParameterHandle(self.handle, self.name, 'attenuation_image')
+        return image
+
+    def set_resolution_model(self, collimator_sigma_0_in_mm, collimator_slope_in_mm, full_3D = True):
+        '''
+        Set the parameters for the depth-dependent resolution model
+
+        The detector and collimator blurring is modelled as a Gaussian with sigma dependent on the
+        distance from the collimator.
+
+        sigma_at_depth = collimator_slope * depth_in_mm + collimator sigma 0
+
+        Set slope and sigma_0 to zero to avoid resolution modelling.
+
+        You have to call set_up() after this.
+        '''
+        try_calling(pystir.cSTIR_SPECTUBMatrixSetResolution(self.handle, collimator_sigma_0_in_mm, collimator_slope_in_mm, full_3D))
 
 class AcquisitionData(DataContainer):
     '''Class for PET acquisition data.'''
@@ -1265,8 +1327,7 @@ class AcquisitionModelUsingMatrix(AcquisitionModel):
         check_status(self.handle)
         if matrix is None:
             matrix = RayTracingMatrix()
-        assert_validity(matrix, RayTracingMatrix)
-        parms.set_parameter(self.handle, self.name, 'matrix', matrix.handle)
+        self.set_matrix(matrix)
     def __del__(self):
         if self.handle is not None:
             pyiutil.deleteDataHandle(self.handle)
@@ -1276,7 +1337,10 @@ class AcquisitionModelUsingMatrix(AcquisitionModel):
         matrix:  a matrix object to represent G in acquisition model.
         '''
         # TODO will need to allow for different matrices here
-        assert_validity(matrix, RayTracingMatrix)
+        try:
+            assert_validity(matrix, SPECTUBMatrix)
+        except:
+            assert_validity(matrix, RayTracingMatrix)
         parms.set_parameter(self.handle, self.name, 'matrix', matrix.handle)
 ##    def get_matrix(self):
 ##        '''